backend/requirements.txt
django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
pillow==12.2.0
gunicorn==23.0.0
whitenoise==6.11.0
dj-database-url==3.0.1
